Launch tmux on terminal startup, with tricks
Launch tmux when bash starts, as long as not in vscode and there is a
display, so it doesn't launch in a tty. It creates a new base session if
there isn't one, but if there is, it adds a new session in the same
group. Being in the same group means that it has access to the same
windows, but can operate independently